third_party/blink/web_tests/fast/forms/datetimelocal/datetimelocal-input-type-expected.txt

Tests for writing and reading .type property of HTMLInputElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.type for "datetime-local" is correctly "datetime-local".
PASS input.type for "DATETIME-LOCAL" is correctly "datetime-local".
PASS input.type for " datetime-local " is correctly "text".
PASS input.type for "datetimelocal" is correctly "text".
PASS input.type for "datetime_local" is correctly "text".
PASS successfullyParsed is true

TEST COMPLETE
